feat(GhostBackup.py): add save game handling
Added handlers for counting saved games as well as establishing a max amount. The saved games have been set to overwrite rather than scale infinitely.

BREAKING CHANGE: There is a possibility that the code may bug on initial setup, this has yet to be confirmed. This has not been tested for anything higher than 12 cases, but there is unlikely to be any issues.