fix(GhostBackup.py): fix save interval, backup, restore
Fixed a bug where the save interval was not properly setting in the environment variable. This was due to it being a string rather than an integer. Fixed a bug where save files were only creating new files instead of backing up the actual save file. Fixed a bug where restore was not properly restoring from selected file.